+++ classes/edit_element_form.php
@@ -24,6 +24,10 @@
 
 namespace tool_certificate;
 
+use context;
+use core_form\dynamic_form;
+use moodle_url;
+
 defined('MOODLE_INTERNAL') || die('Direct access to this script is forbidden.');
 
 require_once($CFG->dirroot . '/' . $CFG->admin . '/tool/certificate/includes/colourpicker.php');
@@ -39,7 +43,7 @@ require_once($CFG->dirroot . '/' . $CFG->admin . '/tool/certificate/includes/col
  * @copyright  2013 Mark Nelson <markn@moodle.com>
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class edit_element_form extends modal_form {
+class edit_element_form extends dynamic_form {
 
     /**
      * @var \tool_certificate\element The element object.
@@ -121,24 +125,33 @@ class edit_element_form extends modal_form {
     }
 
     /**
-     * Check if current user has access to this form, otherwise throw exception
+     * Returns context where this form is used
+     * @return context
+     */
+    protected function get_context_for_dynamic_submission(): context {
+        return $this->get_template()->get_context();
+    }
+
+    /**
+     * Checks if current user has access to this form, otherwise throws exception
      *
      * Sometimes permission check may depend on the action and/or id of the entity.
-     * If necessary, form data is available in $this->_ajaxformdata
+     * If necessary, form data is available in $this->_ajaxformdata or
+     * by calling $this->optional_param()
      */
-    public function require_access() {
+    protected function check_access_for_dynamic_submission(): void {
         $this->get_template()->require_can_manage();
     }
 
     /**
-     * Process the form submission
+     * Process the form submission, used if form was submitted via AJAX
      *
      * This method can return scalar values or arrays that can be json-encoded, they will be passed to the caller JS.
      *
-     * @param \stdClass $data
      * @return \stdClass
      */
-    public function process(\stdClass $data) {
+    public function process_dynamic_submission(): \stdClass {
+        $data = $this->get_data();
         $this->get_element()->save_form_data($data);
         $data = $this->get_element()->to_record();
         // TODO use exporter instead.
@@ -153,7 +166,20 @@ class edit_element_form extends modal_form {
      * Can be overridden to retrieve existing values from db by entity id and also
      * to preprocess editor and filemanager elements
      */
-    public function set_data_for_modal() {
+    public function set_data_for_dynamic_submission(): void {
         $this->set_data($this->get_element()->prepare_data_for_form());
+    }
+
+    /**
+     * Returns url to set in $PAGE->set_url() when form is being rendered or submitted via AJAX
+     *
+     * This is used in the form elements sensitive to the page url, such as Atto autosave in 'editor'
+     *
+     * @return moodle_url
+     */
+    protected function get_page_url_for_dynamic_submission(): moodle_url {
+        return new moodle_url('/admin/tool/certificate/template.php', [
+            'id' => $this->get_template()->get_id(),
+        ]);
     }
 }