+++ element/text/classes/element.php
@@ -66,14 +66,7 @@ class element extends \tool_certificate\element {
      * @param \stdClass $issue the issue we are rendering
      */
     public function render($pdf, $preview, $user, $issue) {
-        global $DB;
-
-        $context = \context_system::instance();
-        // If the issue was generated in a course, use course context instead.
-        if (isset($issue->courseid) && $DB->record_exists('course', ['id' => $issue->courseid])) {
-            $context = \context_course::instance($issue->courseid);
-        }
-        $text = format_text($this->get_data(), FORMAT_HTML, ['context' => $context]);
+        $text = \tool_certificate\element_helper::format_text($this->get_data(), $issue->courseid ?? 0);
         \tool_certificate\element_helper::render_content($pdf, $this, $text);
     }
 
@@ -86,7 +79,7 @@ class element extends \tool_certificate\element {
      * @return string the html
      */
     public function render_html() {
-        $text = format_text($this->get_data(), FORMAT_HTML, ['context' => \context_system::instance()]);
+        $text = \tool_certificate\element_helper::format_text($this->get_data());
         return \tool_certificate\element_helper::render_html_content($this, $text);
     }